+++ plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.23](https://github.com/nativescript-community/ui-chart/compare/v1.2.22...v1.2.23) (2022-11-14)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+
+
+
+
+## [1.2.22](https://github.com/nativescript-community/ui-chart/compare/v1.2.21...v1.2.22) (2022-11-13)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+
+
+
+
 ## [1.2.21](https://github.com/nativescript-community/ui-chart/compare/v1.2.20...v1.2.21) (2022-10-26)
 
 **Note:** Version bump only for package @nativescript-community/ui-chart